+++ tests/test_compensation/test_design_spaces.py
@@ -146,7 +146,6 @@ def simulation_outputs(
     return ref_simulation_output, fix_simulation_output
 
 
-@pytest.mark.implementation
 class TestDesignSpaceImplementation:
 
     _w_kin_tol = 1e-3